+++ js/modules/hist/TF2Painter.mjs
@@ -137,19 +137,27 @@ class TF2Painter extends TH2Painter {
       if (this._use_saved_points) {
          const npx = Math.round(func.fSave[nsave-2]),
                npy = Math.round(func.fSave[nsave-1]),
-               dx = (func.fSave[nsave-5] - func.fSave[nsave-6]) / (npx - 1),
-               dy = (func.fSave[nsave-3] - func.fSave[nsave-4]) / (npy - 1);
+               xmin = func.fSave[nsave-6], xmax = func.fSave[nsave-5],
+               ymin = func.fSave[nsave-4], ymax = func.fSave[nsave-3],
+               epsilon = 1e-10,
+               dx = (func.fXmax - func.fXmin) / Math.max(1, func.fNpx),
+               dy = (func.fYmax - func.fYmin) / Math.max(1, func.fNpy),
+               bad_save_buffer = (Math.abs(xmin - func.fXmin - 0.5*dx) < epsilon) && (Math.abs(func.fXmax - 0.5*dx - xmax) < epsilon) &&
+                                 (Math.abs(ymin - func.fYmin - 0.5*dy) < epsilon) && (Math.abs(func.fYmax - 0.5*dy - ymax) < epsilon),
+               nbinsx = bad_save_buffer ? npx - 1 : npx,
+               nbinsy = bad_save_buffer ? npy - 1 : npy;
 
-         ensureBins(npx+1, npy+1);
-         hist.fXaxis.fXmin = func.fSave[nsave-6] - dx/2;
-         hist.fXaxis.fXmax = func.fSave[nsave-5] + dx/2;
-         hist.fYaxis.fXmin = func.fSave[nsave-4] - dy/2;
-         hist.fYaxis.fXmax = func.fSave[nsave-3] + dy/2;
+         ensureBins(nbinsx, nbinsy);
+         hist.fXaxis.fXmin = xmin;
+         hist.fXaxis.fXmax = xmax;
+         hist.fYaxis.fXmin = ymin;
+         hist.fYaxis.fXmax = ymax;
 
          for (let k = 0, j = 0; j <= npy; ++j) {
             for (let i = 0; i <= npx; ++i) {
                const z = func.fSave[k++];
-               hist.setBinContent(hist.getBin(i+1, j+1), Number.isFinite(z) ? z : 0);
+               if ((j < nbinsy) && (i < nbinsx))
+                  hist.setBinContent(hist.getBin(i+1, j+1), Number.isFinite(z) ? z : 0);
             }
          }
       }
@@ -178,15 +186,10 @@ class TF2Painter extends TH2Painter {
       const func = this.$func, nsave = func?.fSave.length ?? 0;
 
       if (nsave > 6 && this._use_saved_points) {
-         const npx = Math.round(func.fSave[nsave-2]),
-             npy = Math.round(func.fSave[nsave-1]),
-             dx = (func.fSave[nsave-5] - func.fSave[nsave-6]) / (npx - 1),
-             dy = (func.fSave[nsave-3] - func.fSave[nsave-4]) / (npy - 1);
-
-         this.xmin = Math.min(this.xmin, func.fSave[nsave-6] - dx/2);
-         this.xmax = Math.max(this.xmax, func.fSave[nsave-5] + dx/2);
-         this.ymin = Math.min(this.ymin, func.fSave[nsave-4] - dy/2);
-         this.ymax = Math.max(this.ymax, func.fSave[nsave-3] + dy/2);
+         this.xmin = Math.min(this.xmin, func.fSave[nsave-6]);
+         this.xmax = Math.max(this.xmax, func.fSave[nsave-5]);
+         this.ymin = Math.min(this.ymin, func.fSave[nsave-4]);
+         this.ymax = Math.max(this.ymax, func.fSave[nsave-3]);
       }
       if (func) {
          this.xmin = Math.min(this.xmin, func.fXmin);