+++ js/modules/hist/hist3d.mjs
@@ -387,9 +387,10 @@ function create3DControl(fp) {
 
    fp.control.processMouseMove = function(intersects) {
       let tip = null, mesh = null, zoom_mesh = null;
+      const handle_tooltip = frame_painter.isTooltipAllowed();
 
       for (let i = 0; i < intersects.length; ++i) {
-         if (isFunc(intersects[i].object?.tooltip)) {
+         if (handle_tooltip && isFunc(intersects[i].object?.tooltip)) {
             tip = intersects[i].object.tooltip(intersects[i]);
             if (tip) { mesh = intersects[i].object; break; }
          } else if (intersects[i].object?.zoom && !zoom_mesh)
@@ -570,6 +571,35 @@ function change3DCamera(orthographic) {
       create3DControl(this);
 
    this.render3D();
+}
+
+/** @summary Add 3D mesh to frame painter
+  * @private */
+function add3DMesh(mesh, painter, the_only) {
+   if (!mesh)
+      return;
+   if (!this.toplevel)
+      return console.error('3D objects are not yet created in the frame');
+   if (painter && the_only)
+      this.remove3DMeshes(painter);
+   this.toplevel.add(mesh);
+   mesh._painter = painter;
+}
+
+/** @summary Remove 3D meshed for specified painter
+  * @private */
+function remove3DMeshes(painter) {
+   if (!painter || !this.toplevel)
+      return;
+   let i = this.toplevel.children.length;
+
+   while (i > 0) {
+      const mesh = this.toplevel.children[--i];
+      if (mesh._painter === painter) {
+         this.toplevel.remove(mesh);
+         disposeThreejsObject(mesh);
+      }
+   }
 }
 
 
@@ -1427,7 +1457,7 @@ function convert3DtoPadNDC(x, y, z) {
 /** @summary Assign 3D methods for frame painter
   * @private */
 function assignFrame3DMethods(fpainter) {
-   Object.assign(fpainter, { create3DScene, render3D, resize3D, change3DCamera, highlightBin3D, set3DOptions, drawXYZ, convert3DtoPadNDC });
+   Object.assign(fpainter, { create3DScene, add3DMesh, remove3DMeshes, render3D, resize3D, change3DCamera, highlightBin3D, set3DOptions, drawXYZ, convert3DtoPadNDC });
 }
 
 /** @summary Draw histograms in 3D mode
@@ -1671,7 +1701,7 @@ function drawBinsLego(painter, is_v7 = false) {
          return tip;
       };
 
-      main.toplevel.add(mesh);
+      main.add3DMesh(mesh);
 
       if (num2vertices > 0) {
          const geom2 = createLegoGeom(painter, pos2, norm2),
@@ -1687,7 +1717,7 @@ function drawBinsLego(painter, is_v7 = false) {
          mesh2.baseline = mesh.baseline;
          mesh2.tip_color = mesh.tip_color;
 
-         main.toplevel.add(mesh2);
+         main.add3DMesh(mesh2);
       }
    }
 
@@ -1780,7 +1810,7 @@ function drawBinsLego(painter, is_v7 = false) {
    }
    */
 
-   main.toplevel.add(line);
+   main.add3DMesh(line);
 }
 
 /** @summary Draw TH2 histogram in error mode
@@ -1887,7 +1917,7 @@ function drawBinsError3D(painter, is_v7 = false) {
        return tip;
     };
 
-    main.toplevel.add(line);
+    main.add3DMesh(line);
 }
 
 /** @summary Draw TH2 as 3D contour plot
@@ -1920,7 +1950,7 @@ function drawBinsContour3D(painter, realz = false, is_v7 = false) {
    );
 
    const lines = createLineSegments(pnts, create3DLineMaterial(painter, is_v7 ? 'line_' : histo));
-   main.toplevel.add(lines);
+   main.add3DMesh(lines);
 }
 
 /** @summary Draw TH2 histograms in surf mode
@@ -2039,7 +2069,7 @@ function drawBinsSurf3D(painter, is_v7 = false) {
 
       const mesh = new Mesh(geometry, material);
 
-      main.toplevel.add(mesh);
+      main.add3DMesh(mesh);
 
       mesh.painter = painter; // to let use it with context menu
    }, (isgrid, lpos) => {
@@ -2056,7 +2086,7 @@ function drawBinsSurf3D(painter, is_v7 = false) {
 
       const line = createLineSegments(convertLegoBuf(painter, lpos, handle.i2 - handle.i1, handle.j2 - handle.j1), material);
       line.painter = painter;
-      main.toplevel.add(line);
+      main.add3DMesh(line);
    });
 
    if (painter.options.Surf === 17)
@@ -2119,7 +2149,7 @@ function drawBinsSurf3D(painter, is_v7 = false) {
                    material = new MeshBasicMaterial(getMaterialArgs(palette.getColor(colindx), { side: DoubleSide, opacity: 0.5, vertexColors: false })),
                    mesh = new Mesh(geometry, material);
              mesh.painter = painter;
-             main.toplevel.add(mesh);
+             main.add3DMesh(mesh);
          }
       );
    }